Add play function to program
Added the play function to the program,this function has two parameters

The first parameter is the player choice and the second parameter is the computer choice,
the structure of the code inside the play function is a switch staement in tandem with if and if else statements.
The switch staements code are executed based upon whether the computer chooses "Rock", "Paper" or "Scissors".
If and else if statements inside the switch statements are used to compare the computer's choice with the choice of the player.