Add game function
The game function plays 5 rounds of rock paper scissors

The game function plays 5 rounds and outputs the wins or losses the player has recived during
the game at the end. The function was created to play 5 rounds while ate the same time, incopparinting
the amount of wins or losses stored in variables.