The div element in the progam displays the result of the round
Fix div element to display result

In the program a div element was created as a dom method in the program.

The div element displays the result of the choice chosen by the user compared to that of the computer.
In the play function the console.log statements were all converted to dom methods.
A single paragrph element was created under the container dom method, and the same p element is updated based
upon the result of the round. All the console log statements inside of the play function were converted to using the p.textContent property;