Add Play Again button
A play again button ws created

In the checkWinner function a play again button was created.

The checkWinner function serves two purposes.
1. The first purpose is in it's name, so it basically checks the winner,
2. The second purpose it surves is to restart the program by reseting the win and loss values.
Also it is important to mention that in this function a play agian butoon is created and added to the
container. When the button is clicked it then resets the values of win and loss and then disapears.
This is implented in each of the differnt cases, whether the user wins, or the cpu ,and if there is a draw.